fix: rename Catalog to CatalogContainer
